@@ -8,33 +8,32 @@
     "run-build": "npm run build && next start",
     "deploy": "npm run build && gh-pages -t -d out",
     "build": "next build && next export && npm run generate-sitemap && npm run create-404",
-    "generate-sitemap": "node generate-sitemap",
+    "generate-sitemap": "node generate-sitemap.mjs",
     "create-404": "move-file out/404/index.html out/404.html",
     "prettier": "prettier --write ."
   },
   "dependencies": {
-    "@types/commonmark": "^0.27.4",
-    "@types/github-slugger": "^1.3.0",
-    "@types/styled-components": "^5.1.4",
-    "axios": "^0.21.1",
+    "axios": "^0.24.0",
     "commonmark": "^0.29.2",
-    "github-slugger": "^1.3.0",
+    "github-slugger": "^1.4.0",
     "move-file-cli": "^2.0.0",
-    "next": "10.0.1",
-    "polished": "^4.0.3",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "next": "12.0.4",
+    "polished": "^4.1.3",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-markdown": "^5.0.2",
     "remark-gfm": "^1.0.0",
-    "sass": "^1.29.0",
-    "styled-components": "^5.2.1"
+    "styled-components": "^5.3.3"
   },
   "devDependencies": {
-    "@types/node": "^14.14.7",
-    "@types/react": "^16.9.56",
-    "gh-pages": "^3.1.0",
-    "globby": "^11.0.1",
-    "prettier": "^2.1.2",
-    "typescript": "^4.0.5"
+    "@types/node": "^16.11.9",
+    "@types/react": "^17.0.36",
+    "@types/commonmark": "^0.27.4",
+    "@types/github-slugger": "^1.3.0",
+    "@types/styled-components": "^5.1.15",
+    "gh-pages": "^3.2.3",
+    "globby": "^12.0.2",
+    "prettier": "^2.4.1",
+    "typescript": "^4.5.2"
   }
 }
