@@ -1,14 +1,5 @@
-import { useAmp } from 'next/amp';
-
 export const Analytics: React.FunctionComponent<{ id: string }> = ({ id }) => {
-  const isAmp = useAmp();
-  return isAmp ? (
-    <script
-      async
-      custom-element="amp-analytics"
-      src="https://cdn.ampproject.org/v0/amp-analytics-0.1.js"
-    />
-  ) : (
+  return (
     <>
       <script defer src={`https://www.googletagmanager.com/gtag/js?id=${id}`} />
       <script
