@@ -1,4 +1,4 @@
-import { SmartImage } from '../smart-image';
+import Image from 'next/image';
 
 import {
   ItemWrapper,
@@ -16,20 +16,37 @@ export type CardGridItemProps = {
   imageUrl?: string;
 };
 
+const stringToNumber = function (str: string, seed = 3) {
+  let h1 = 0xdeadbeef ^ seed,
+    h2 = 0x41c6ce57 ^ seed;
+  for (let i = 0, ch; i < str.length; i++) {
+    ch = str.charCodeAt(i);
+    h1 = Math.imul(h1 ^ ch, 2654435761);
+    h2 = Math.imul(h2 ^ ch, 1597334677);
+  }
+  h1 =
+    Math.imul(h1 ^ (h1 >>> 16), 2246822507) ^
+    Math.imul(h2 ^ (h2 >>> 13), 3266489909);
+  h2 =
+    Math.imul(h2 ^ (h2 >>> 16), 2246822507) ^
+    Math.imul(h1 ^ (h1 >>> 13), 3266489909);
+  return 4294967296 * (2097151 & h2) + (h1 >>> 0);
+};
+
 export const CardGridItem: React.FunctionComponent<CardGridItemProps> = ({
   title,
   imageUrl,
   description,
 }) => (
   <ItemWrapper>
-    <ImageWrapper hue={imageUrl ? 0 : Math.random()}>
+    <ImageWrapper hue={imageUrl ? 0 : stringToNumber(title)}>
       {!imageUrl && <PlaceholderText>{title}</PlaceholderText>}
-      <SmartImage
-        layout="fixed"
+      <Image
+        layout="intrinsic"
         alt={title}
         src={imageUrl || '/images/placeholder.jpg'}
-        height="100px"
-        width="280px"
+        height={200}
+        width={560}
       />
     </ImageWrapper>
     <TextWrapper>
