@@ -1,6 +1,5 @@
-import { useAmp } from 'next/amp';
+import Image from 'next/image';
 import { ImageMap } from '../../helpers';
-import { SmartImage } from '../smart-image';
 
 type Props = {
   src: string;
@@ -11,16 +10,16 @@ type Props = {
 export const ImageRenderer = (
   externalImages: ImageMap
 ): React.FunctionComponent<Props> => ({ src, alt, node, ...props }) => {
-  const isAmp = useAmp();
-
-  return (!isAmp || externalImages[src]) ? (
-    <SmartImage
+  return externalImages[src] ? (
+    <Image
       alt={alt}
       src={externalImages[src] ?? src}
-      height={isAmp ? 100 : undefined}
-      width={isAmp ? 1000 : undefined}
+      height={15}
+      width={60}
       layout="responsive"
       {...props}
     />
-  ) : null;
+  ) : (
+    <img src={src} />
+  );
 };
